Added userID as a requirement for two requests.

+++ api/server.js
@@ -55,7 +55,7 @@ app.post('/users/', function (req, res) {
 
 // Get a list of all contacts for a user
 app.get('/contacts/', function (req, res) {
-  var query = dbConnection.query('SELECT * FROM Contacts WHERE firstName = "' + req.query.firstName + '" AND lastName = "' + req.query.lastName + '"', function (error, results, fields) {
+  var query = dbConnection.query('SELECT * FROM Contacts WHERE userID = "' + req.query.userID + '" AND firstName = "' + req.query.firstName + '" AND lastName = "' + req.query.lastName + '"', function (error, results, fields) {
     console.log(error)
     if (error) error = 1
     res.send(JSON.stringify({'status': 200, 'error': error, 'response': results}))
@@ -64,7 +64,7 @@ app.get('/contacts/', function (req, res) {
 
 // Delete Contacts entry
 app.delete('/contacts/', function (req, res) {
-  var query = dbConnection.query('DELETE FROM Contacts WHERE firstName = "' + req.body.firstName + '" AND lastName = "' + req.body.lastName + '"', function (error, results, fields) {
+  var query = dbConnection.query('DELETE FROM Contacts WHERE userID = "' + req.body.userID + '" AND firstName = "' + req.body.firstName + '" AND lastName = "' + req.body.lastName + '"', function (error, results, fields) {
     console.log(error)
     if (error) error = 1
     res.send(JSON.stringify({'status': 200, 'error': error, 'response': results}))
